Give focus to first field in a Document field, refs ERM-575
