+++ lib/DocumentsFieldArray/DocumentsFieldArray.js
@@ -1,39 +1,32 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { get } from 'lodash';
 import { FormattedMessage } from 'react-intl';
 import { Field } from 'react-final-form';
 
 import {
   Button,
-  Col,
   Layout,
-  Row,
-  Select,
-  TextArea,
-  TextField,
 } from '@folio/stripes/components';
 
 import EditCard from '../EditCard';
-import FileUploaderField from '../FileUploaderField';
 import withKiwtFieldArray from '../withKiwtFieldArray';
-import { composeValidators } from '../validators';
+import DocumentField from './DocumentField';
 
 class DocumentsFieldArray extends React.Component {
   static propTypes = {
     addDocBtnLabel: PropTypes.node,
     deleteButtonTooltipText: PropTypes.node,
-    onDownloadFile: PropTypes.func,
-    onUploadFile: PropTypes.func,
+    documentCategories: PropTypes.arrayOf(PropTypes.shape({
+      label: PropTypes.string,
+      value: PropTypes.string,
+    })),
     isEmptyMessage: PropTypes.node,
     items: PropTypes.arrayOf(PropTypes.object),
     name: PropTypes.string.isRequired,
     onAddField: PropTypes.func.isRequired,
     onDeleteField: PropTypes.func.isRequired,
-    documentCategories: PropTypes.arrayOf(PropTypes.shape({
-      label: PropTypes.string,
-      value: PropTypes.string,
-    })),
+    onDownloadFile: PropTypes.func,
+    onUploadFile: PropTypes.func,
   }
 
   static defaultProps = {
@@ -41,68 +34,15 @@ class DocumentsFieldArray extends React.Component {
     isEmptyMessage: <FormattedMessage id="stripes-erm-components.doc.noDocs" />,
   }
 
-  validateDocIsSpecified = (value, allValues, meta) => {
-    if (meta) {
-      const index = parseInt(/\[([0-9]*)\]/.exec(meta.name)[1], 10);
-      const { fileUpload, location, name, url } = get(allValues, [this.props.name, index], {});
-      if (name && (!fileUpload && !location && !url)) {
-        return <FormattedMessage id="stripes-erm-components.doc.error.docsMustHaveLocationOrURL" />;
-      }
-    }
-
-    return undefined;
-  }
-
-  validateURLIsValid = (value) => {
-    if (value) {
-      try {
-        // Test if the URL is valid
-        new URL(value); // eslint-disable-line no-new
-      } catch (_) {
-        return <FormattedMessage id="stripes-erm-components.doc.error.invalidURL" />;
-      }
-    }
-
-    return undefined;
-  }
-
-  validateRequired = (value) => (
-    !value ? <FormattedMessage id="stripes-core.label.missingRequiredField" /> : undefined
-  )
-
-  renderCategory = (i) => {
-    const { documentCategories, name } = this.props;
-
-    if (!get(documentCategories, 'length')) return null;
-
-    return (
-      <Row>
-        <Col xs={12}>
-          <Field
-            component={Select}
-            data-test-document-field-category
-            dataOptions={[
-              { value: '', label: '' },
-              ...documentCategories,
-            ]}
-            id={`${name}-category-${i}`}
-            label={<FormattedMessage id="stripes-erm-components.doc.category" />}
-            name={`${name}[${i}].atType`}
-            parse={v => v} // Lets us send an empty string instead of `undefined`
-          />
-        </Col>
-      </Row>
-    );
-  }
-
   renderDocs = () => {
     const {
       deleteButtonTooltipText,
-      onDownloadFile,
-      onUploadFile,
+      documentCategories,
       items,
       name,
-      onDeleteField
+      onDeleteField,
+      onDownloadFile,
+      onUploadFile
     } = this.props;
 
     return items.map((doc, i) => (
@@ -117,79 +57,14 @@ class DocumentsFieldArray extends React.Component {
         key={i}
         onDelete={() => onDeleteField(i, doc)}
       >
-        <Row>
-          <Col xs={12} md={onUploadFile ? 6 : 12}>
-            <Row>
-              <Col xs={12}>
-                <Field
-                  data-test-document-field-name
-                  component={TextField}
-                  id={`${name}-name-${i}`}
-                  label={<FormattedMessage id="stripes-erm-components.doc.name" />}
-                  name={`${name}[${i}].name`}
-                  required
-                  validate={this.validateRequired}
-                />
-              </Col>
-            </Row>
-            {this.renderCategory(i)}
-            <Row>
-              <Col xs={12}>
-                <Field
-                  data-test-document-field-note
-                  component={TextArea}
-                  id={`${name}-note-${i}`}
-                  label={<FormattedMessage id="stripes-erm-components.doc.note" />}
-                  name={`${name}[${i}].note`}
-                  parse={v => v} // Lets us send an empty string instead of `undefined`
-                />
-              </Col>
-            </Row>
-          </Col>
-          {onUploadFile &&
-            <Col xs={12} md={6}>
-              <Field
-                component={FileUploaderField}
-                data-test-document-field-file
-                id={`${name}-file-${i}`}
-                label={<FormattedMessage id="stripes-erm-components.doc.file" />}
-                name={`${name}[${i}].fileUpload`}
-                onDownloadFile={onDownloadFile}
-                onUploadFile={onUploadFile}
-                validate={this.validateDocIsSpecified}
-              />
-            </Col>
-          }
-        </Row>
-        <Row>
-          <Col xs={12}>
-            <Field
-              data-test-document-field-location
-              component={TextField}
-              id={`${name}-location-${i}`}
-              label={<FormattedMessage id="stripes-erm-components.doc.location" />}
-              name={`${name}[${i}].location`}
-              parse={v => v} // Lets us send an empty string instead of `undefined`
-              validate={this.validateDocIsSpecified}
-            />
-          </Col>
-        </Row>
-        <Row>
-          <Col xs={12}>
-            <Field
-              component={TextField}
-              data-test-document-field-url
-              id={`${name}-url-${i}`}
-              label={<FormattedMessage id="stripes-erm-components.doc.url" />}
-              name={`${name}[${i}].url`}
-              parse={v => v} // Lets us send an empty string instead of `undefined`
-              validate={composeValidators(
-                this.validateDocIsSpecified,
-                this.validateURLIsValid,
-              )}
-            />
-          </Col>
-        </Row>
+        <Field
+          component={DocumentField}
+          documentCategories={documentCategories}
+          index={i}
+          name={`${name}`}
+          onDownloadFile={onDownloadFile}
+          onUploadFile={onUploadFile}
+        />
       </EditCard>
     ));
   }
