+++ lib/DocumentsFieldArray/tests/DocumentsFieldArray-test.js
@@ -14,8 +14,8 @@ const { expect, spy } = chai;
 const uploadedFile = { id: '1', name: 'Test File', modified: Date.now() };
 const onDownloadFile = spy();
 const onUploadFile = spy(() => {
-  const blob = new Blob([JSON.stringify(uploadedFile, null, 2)], { type : 'application/json' });
-  const init = { 'status' : 200, 'statusText': 'OK' };
+  const blob = new Blob([JSON.stringify(uploadedFile, null, 2)], { type: 'application/json' });
+  const init = { 'status': 200, 'statusText': 'OK' };
   const myResponse = new Response(blob, init);
   return Promise.resolve(myResponse);
 });
@@ -25,7 +25,7 @@ const isEmptyMessage = 'No documents';
 
 const interactor = new DocumentsFieldArrayInteractor();
 
-describe('DocumentsFieldArray', () => {
+describe.only('DocumentsFieldArray', () => {
   beforeEach(async () => {
     await mountWithContext(
       <TestForm>
